+++ test/test_stages.py
@@ -44,12 +44,11 @@ def test_IFStage(sim):
 
     # SW a0,-20(s0) = SW, x10, -20(x8)
     fetch.imem.writeRequest(0, 0xfea42623, 4)
-    fetch.imem._tick()
+    sim.step()
     fetch.npc_i.write(0x00000000)
 
-    sim._process_queue()
-    RegBase.tick()
-    sim._process_queue()
+    sim.step()
+    sim.step()
 
     out = fetch.IFID_o.read()
     assert out.inst == 0xfea42623
@@ -138,27 +137,27 @@ class TestIDStage:
         # No exception for valid instruction
         inst = 0x23 # store
         dec.IFID_i.write(IFID_t(inst, pc))
-        sim._process_queue()
+        sim.step()
 
         ## Inst[1:0] != 2'b11
         inst = 0x10
         pc += 1
         dec.IFID_i.write(IFID_t(inst, pc))
         with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X} detected."):
-            sim._process_queue()
+            sim.step()
 
         ## Unsupported RV32base Opcodes
         inst = 0x1F # opcode = 0011111
         pc += 1
         dec.IFID_i.write(IFID_t(inst, pc))
         with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X}: unknown opcode"):
-            sim._process_queue()
+            sim.step()
 
         inst = 0x73 # opcode = 1110011
         pc += 1
         dec.IFID_i.write(IFID_t(inst, pc))
         with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X}: unknown opcode"):
-            sim._process_queue()
+            sim.step()
 
         ## Illegal combinations of funct3, funct7
 
@@ -168,13 +167,13 @@ class TestIDStage:
         pc += 1
         dec.IFID_i.write(IFID_t(inst, pc))
         with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X} detected."):
-            sim._process_queue()
+            sim.step()
         # If funct3 == 5 => funct7 == {0, 0100000}
         inst = 0xc0005013 # funct7 = 1100000
         pc += 1
         dec.IFID_i.write(IFID_t(inst, pc))
         with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X} detected."):
-            sim._process_queue()
+            sim.step()
 
         # ADD - AND -> opcode = 0110011
         # If funct7 != {0, 0100000} -> illegal
@@ -182,20 +181,20 @@ class TestIDStage:
         pc += 1
         dec.IFID_i.write(IFID_t(inst, pc))
         with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X} detected."):
-            sim._process_queue()
+            sim.step()
         # If funct7 == 0100000 => funct3 == {0, 5}
         inst = 0x40002033 # funct3 = 2
         pc += 1
         dec.IFID_i.write(IFID_t(inst, pc))
         with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X} detected."):
-            sim._process_queue()
+            sim.step()
 
         # JALR -> opcode = 1100111 => funct3 == 0
         inst = 0x00005067 # funct3 = 5
         pc += 1
         dec.IFID_i.write(IFID_t(inst, pc))
         with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X} detected."):
-            sim._process_queue()
+            sim.step()
 
         # BEQ - BGEU -> opcode = 1100011 => funct3 = {0,1,4,5,6,7}
         funct3 = [2,3]
@@ -204,7 +203,7 @@ class TestIDStage:
             inst = 0x63 | (f3 << 12)
             dec.IFID_i.write(IFID_t(inst, pc))
             with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X} detected."):
-                sim._process_queue()
+                sim.step()
 
         # LB - LHU -> opcode = 0000011 => funct3 = {0,1,2,4,5}
         funct3 = [3,6,7]
@@ -213,7 +212,7 @@ class TestIDStage:
             inst = 0x3 | (f3 << 12)
             dec.IFID_i.write(IFID_t(inst, pc))
             with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X} detected."):
-                sim._process_queue()
+                sim.step()
 
         # SB, SH, SW -> opcode = 0100011 => funct3 = {0,1,2}
         funct3 = [3,4,5,6,7]
@@ -222,7 +221,7 @@ class TestIDStage:
             inst = 0x23 | (f3 << 12)
             dec.IFID_i.write(IFID_t(inst, pc))
             with pytest.raises(Exception, match = f"IDStage: Illegal instruction @ PC = 0x{pc:08X} detected."):
-                sim._process_queue()
+                sim.step()
 
     def test_IDStage(self, sim):
         regf = Regfile()
@@ -231,13 +230,13 @@ class TestIDStage:
         # SW a0,-20(s0) = SW, x10, -20(x8) (x8=rs1, x10=rs2)
         # Write some values into the relevant registers
         regf.writeRequest(8, 0x80000000)
-        regf._tick()
+        sim.step()
         regf.writeRequest(10, 42)
-        regf._tick()
+        sim.step()
 
         # Set inputs
         decode.IFID_i.write(IFID_t(0xfea42623, 0x80000004))
-        sim._process_queue()
+        sim.step()
 
         # Validate outputs
         out = decode.IDEX_o.read()
@@ -254,7 +253,7 @@ class TestIDStage:
         # Test instruction with register write
         # ADDI x0, x0, 0
         decode.IFID_i.write(IFID_t(0x00000013, 0x80000004))
-        sim._process_queue()
+        sim.step()
         out = decode.IDEX_o.read()
         assert out.rs1 == 0
         assert out.rs2 == 0
@@ -270,12 +269,12 @@ class TestIDStage:
         # Test instruction with funct7 output
         # SUB x14, x7, x5
         regf.writeRequest(7, 43)
-        regf._tick()
+        sim.step()
         regf.writeRequest(5, 12)
-        regf._tick()
+        sim.step()
 
         decode.IFID_i.write(IFID_t(0x40538733, 0x80000004))
-        sim._process_queue()
+        sim.step()
         out = decode.IDEX_o.read()
         assert out.rs1 == 43
         assert out.rs2 == 12
@@ -299,9 +298,9 @@ class TestIDStage:
         # Test LOAD
         # LW x15, x8, 0x456
         regf.writeRequest(8, 0x40000000)
-        regf._tick()
+        sim.step()
         decode.IFID_i.write(IFID_t(0x45642783, 0x80000004))
-        sim._process_queue()
+        sim.step()
         out = decode.IDEX_o.read()
         assert out.rs1 == 0x40000000
         assert out.imm == 0x456
@@ -328,7 +327,7 @@ class TestEXStage:
 
         ex.IDEX_i.write(IDEX_t(rd=1, we=1, wb_sel=2, rs2=23, mem=1, funct3=5))
 
-        sim._process_queue()
+        sim.step()
 
         out = ex.EXMEM_o.read()
         assert out.rd == 1
@@ -792,7 +791,7 @@ class TestEXStage:
             wb_sel=0,
             opcode=0b01101
         ))
-        sim._process_queue()
+        sim.step()
         out = ex.EXMEM_o.read()
         assert out.take_branch == False
         assert out.alu_res == 0xaffe<<12
@@ -811,7 +810,7 @@ class TestEXStage:
             wb_sel=0,
             opcode=0b00101
         ))
-        sim._process_queue()
+        sim.step()
         out = ex.EXMEM_o.read()
         assert out.take_branch == False
         assert out.alu_res == 0x8AFFE000
@@ -830,7 +829,7 @@ class TestEXStage:
             wb_sel=1,
             opcode=0b11011
         ))
-        sim._process_queue()
+        sim.step()
         out = ex.EXMEM_o.read()
         assert out.take_branch == True
         assert out.alu_res == 0x8005F514
@@ -850,7 +849,7 @@ class TestEXStage:
             wb_sel=1,
             opcode=0b11001
         ))
-        sim._process_queue()
+        sim.step()
         out = ex.EXMEM_o.read()
         assert out.take_branch == True
         assert out.alu_res == 0x4600
@@ -875,7 +874,7 @@ class TestEXStage:
             opcode=0b11011
         ))
         with pytest.raises(Exception, match = f"Target instruction address misaligned exception at PC = 0x{pc:08X}"):
-            sim._process_queue()
+            sim.step()
         pc += 4
 
         # JALR x13, rs1, 0xA8A
@@ -888,7 +887,7 @@ class TestEXStage:
             opcode=0b11001
         ))
         with pytest.raises(Exception, match = f"Target instruction address misaligned exception at PC = 0x{pc:08X}"):
-            sim._process_queue()
+            sim.step()
         pc += 4
 
         # BEQ
@@ -901,7 +900,7 @@ class TestEXStage:
             opcode=0b11000
         ))
         with pytest.raises(Exception, match = f"Target instruction address misaligned exception at PC = 0x{pc:08X}"):
-            sim._process_queue()
+            sim.step()
         pc += 4
 
         # BNE
@@ -914,7 +913,7 @@ class TestEXStage:
             opcode=0b11000
         ))
         with pytest.raises(Exception, match = f"Target instruction address misaligned exception at PC = 0x{pc:08X}"):
-            sim._process_queue()
+            sim.step()
         pc += 4
 
         # BLT
@@ -927,7 +926,7 @@ class TestEXStage:
             opcode=0b11000
         ))
         with pytest.raises(Exception, match = f"Target instruction address misaligned exception at PC = 0x{pc:08X}"):
-            sim._process_queue()
+            sim.step()
         pc += 4
 
         # BGE
@@ -940,7 +939,7 @@ class TestEXStage:
             opcode=0b11000
         ))
         with pytest.raises(Exception, match = f"Target instruction address misaligned exception at PC = 0x{pc:08X}"):
-            sim._process_queue()
+            sim.step()
         pc += 4
 
         # BLTU
@@ -953,7 +952,7 @@ class TestEXStage:
             opcode=0b11000
         ))
         with pytest.raises(Exception, match = f"Target instruction address misaligned exception at PC = 0x{pc:08X}"):
-            sim._process_queue()
+            sim.step()
         pc += 4
 
         # BGEU
@@ -966,7 +965,7 @@ class TestEXStage:
             opcode=0b11000
         ))
         with pytest.raises(Exception, match = f"Target instruction address misaligned exception at PC = 0x{pc:08X}"):
-            sim._process_queue()
+            sim.step()
         pc += 4
 
         # No exception for not-taken branch
@@ -979,7 +978,7 @@ class TestEXStage:
             funct3=0,
             opcode=0b11000
         ))
-        sim._process_queue()
+        sim.step()
         pc += 4
 
 
@@ -1005,7 +1004,7 @@ class TestMEMStage:
             pc4=0xdeadbeef,
             alu_res=0xaffeaffe
         ))
-        sim._process_queue()
+        sim.step()
         out = mem.MEMWB_o.read()
         assert out.rd == 1
         assert out.we == 1
@@ -1017,9 +1016,9 @@ class TestMEMStage:
         mem = MEMStage(Memory(1024))
         # Load memory
         mem.mem.writeRequest(0, 0xdeadbeef, 4)
-        mem.mem._tick()
+        sim.step()
         mem.mem.writeRequest(4, 0xbade0123, 4)
-        mem.mem._tick()
+        sim.step()
 
         # LB
         mem.EXMEM_i.write(EXMEM_t(
@@ -1027,7 +1026,7 @@ class TestMEMStage:
             alu_res=2, # addr
             funct3=0 #lb
         ))
-        sim._process_queue()
+        sim.step()
         assert mem.MEMWB_o.read().mem_rdata == 0xffffffad
 
         mem.EXMEM_i.write(EXMEM_t(
@@ -1035,7 +1034,7 @@ class TestMEMStage:
             alu_res=5, # addr
             funct3=0 #lb
         ))
-        sim._process_queue()
+        sim.step()
         assert mem.MEMWB_o.read().mem_rdata == 0x00000001 
 
         # LH
@@ -1044,7 +1043,7 @@ class TestMEMStage:
             alu_res=2, # addr
             funct3=1 # lh
         ))
-        sim._process_queue()
+        sim.step()
         assert mem.MEMWB_o.read().mem_rdata == 0xffffdead
 
         mem.EXMEM_i.write(EXMEM_t(
@@ -1052,7 +1051,7 @@ class TestMEMStage:
             alu_res=4, # addr
             funct3=1 # lh
         ))
-        sim._process_queue()
+        sim.step()
         assert mem.MEMWB_o.read().mem_rdata == 0x00000123 
 
         # LW
@@ -1061,7 +1060,7 @@ class TestMEMStage:
             alu_res=0, # addr
             funct3=2 # lw
         ))
-        sim._process_queue()
+        sim.step()
         assert mem.MEMWB_o.read().mem_rdata == 0xdeadbeef
 
         # LBU
@@ -1070,7 +1069,7 @@ class TestMEMStage:
             alu_res=2, # addr
             funct3=4 # lbu
         ))
-        sim._process_queue()
+        sim.step()
         assert mem.MEMWB_o.read().mem_rdata == 0xad
 
         # LHU
@@ -1079,7 +1078,7 @@ class TestMEMStage:
             alu_res=2, # addr
             funct3=5 # lbu
         ))
-        sim._process_queue()
+        sim.step()
         assert mem.MEMWB_o.read().mem_rdata == 0xdead
 
     def test_store(self, sim):
@@ -1092,12 +1091,11 @@ class TestMEMStage:
             rs2=0xabadbabe, # wdata
             funct3=0 # sb
         ))
-        sim._process_queue()
-        mem.mem._tick()
+        sim.step()
         assert mem.mem.read(3, 1) == 0xbe
 
         mem.mem.writeRequest(0, 0, 4)
-        mem.mem._tick()
+        sim.step()
 
         # SH
         mem.EXMEM_i.write(EXMEM_t(
@@ -1106,8 +1104,7 @@ class TestMEMStage:
             rs2=0xabadbabe, # wdata
             funct3=1 # sh
         ))
-        sim._process_queue()
-        mem.mem._tick()
+        sim.step()
         assert mem.mem.read(0, 2) == 0xbabe
 
         # SW
@@ -1117,8 +1114,7 @@ class TestMEMStage:
             rs2=0xabadbabe, # wdata
             funct3=2 # sw
         ))
-        sim._process_queue()
-        mem.mem._tick()
+        sim.step()
         assert mem.mem.read(0, 4) == 0xabadbabe
 
     def test_exception(self, caplog, sim):
@@ -1131,7 +1127,7 @@ class TestMEMStage:
             alu_res=1, # addr
             funct3=1 # lh
         ))
-        sim._process_queue()
+        sim.step()
         assert f"Misaligned load from address 0x00000001" in caplog.text
         caplog.clear()
 
@@ -1141,7 +1137,7 @@ class TestMEMStage:
             alu_res=3, # addr
             funct3=2 # lw
         ))
-        sim._process_queue()
+        sim.step()
         assert f"Misaligned load from address 0x00000003" in caplog.text
         caplog.clear()
 
@@ -1152,7 +1148,7 @@ class TestMEMStage:
             alu_res=1, # addr
             funct3=1 # sh
         ))
-        sim._process_queue()
+        sim.step()
         assert f"Misaligned store to address 0x00000001" in caplog.text
         caplog.clear()
 
@@ -1162,7 +1158,7 @@ class TestMEMStage:
             alu_res=3, # addr
             funct3=2 # sw
         ))
-        sim._process_queue()
+        sim.step()
         assert f"Misaligned store to address 0x00000003" in caplog.text
         caplog.clear()
 
@@ -1188,8 +1184,7 @@ class TestWBStage:
             mem_rdata=0xdeadbeef,
             wb_sel=0
         ))
-        sim._process_queue()
-        wb.regfile._tick()
+        sim.step()
         assert wb.regfile.read(18) == 42
 
         # PC+4 (JAL)
@@ -1201,8 +1196,7 @@ class TestWBStage:
             mem_rdata=0xdeadbeef,
             wb_sel=1
         ))
-        sim._process_queue()
-        wb.regfile._tick()
+        sim.step()
         assert wb.regfile.read(31) == 87
 
         # Memory load
@@ -1214,15 +1208,14 @@ class TestWBStage:
             mem_rdata=0xdeadbeef,
             wb_sel=2
         ))
-        sim._process_queue()
-        wb.regfile._tick()
+        sim.step()
         assert wb.regfile.read(4) == 0xdeadbeef
 
     def test_no_wb(self, sim):
         wb = WBStage(Regfile())
 
         wb.regfile.writeRequest(25, 1234)
-        wb.regfile._tick()
+        sim.step()
 
         val = MEMWB_t(
             we=0,
@@ -1235,22 +1228,19 @@ class TestWBStage:
         # ALU op
         val.wb_sel = 0
         wb.MEMWB_i.write(val)
-        sim._process_queue()
-        wb.regfile._tick()
+        sim.step()
         assert wb.regfile.read(25) == 1234
 
         # PC+4 (JAL)
         val.wb_sel = 1
         wb.MEMWB_i.write(val)
-        sim._process_queue()
-        wb.regfile._tick()
+        sim.step()
         assert wb.regfile.read(25) == 1234
 
         # Memory load
         val.wb_sel = 2
         wb.MEMWB_i.write(val)
-        sim._process_queue()
-        wb.regfile._tick()
+        sim.step()
         assert wb.regfile.read(25) == 1234
 
 # ---------------------------------------
@@ -1269,12 +1259,12 @@ def test_branch_unit(sim):
     bu.pc_i.write(0x80000000)
     bu.take_branch_i.write(False)
     bu.target_i.write(0x40000000)
-    sim._process_queue()
+    sim.step()
     assert bu.npc_o.read() == 0x80000004 
 
     # Test taken branch
     bu.pc_i.write(0x80000000)
     bu.take_branch_i.write(True)
     bu.target_i.write(0x40000000)
-    sim._process_queue()
+    sim.step()
     assert bu.npc_o.read() == 0x40000000